警告取り

+++ project/Source/Item.cpp
@@ -3,9 +3,9 @@
 
 Item::Item()
 {
-	hImage = LoadGraph("data/textures/item.png");
-	kind = rand() % 5;
-	position = VGet(rand() % (SCREEN_WIDTH - 32), rand() % (SCREEN_HEIGHT - 32), 0);
+	hImage   = LoadGraph("data/textures/item.png");
+	kind     = rand() % 5;
+	position = VGet((float)(rand() % (SCREEN_WIDTH - 32)), (float)(rand() % (SCREEN_HEIGHT - 32)), 0.0f);
 }
 
 Item::~Item()
@@ -22,5 +22,5 @@ void Item::Update()
 
 void Item::Draw()
 {
-	DrawRectGraph(position.x, position.y, kind * 36 + 2, 2, 32, 32, hImage, TRUE);
+	DrawRectGraph((int)position.x, (int)position.y, kind * 36 + 2, 2, 32, 32, hImage, TRUE);
 }
